// frontend/src/app/models/korisnik.ts
export default class Korisnik {
    korisnickoIme: string = "";
    lozinka: string = "";
    ime: string = "";
    prezime: string = "";
    pol: string = "";
    adresa: string = "";
    telefon: string = "";
    imejl: string = "";
    profilnaSlika: string = "";
    tip: string = "";
    deaktiviran: string = "";
    
    brojKartice: string = "";
    odobren: string = "";

    firma: string = "";
    blokiran: string = "";
}